requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.2
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.8.0
cycler==0.12.1
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-rest-swagger==2.2.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
fonttools==4.57.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.4
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
